Force the gem to use httpclient
Normally the HTTPI gem allows users of your gem to pick which HTTP
gem they prefer and it is a simple abstraction layer above whatever
adapter they prefer. We want to ensure that HTTPS connections are
reused for several requests, and the best way to do that is to
explicitly require the httpclient gem. This might mean more
dependencies for projects that use this gem but it should also
result in greater performance.