Convert structs to full-fledged classes
Turn the Document, DocumentType and DocumentwithContent into full-fledged Ruby classes instead of Struct declarations. The main point of this exercise is to define methods for parsing from XML, so that logic for loading Documents from XML nodes is not duplicated in several places like it is now. It also makes the unit tests simple.