Add parsed Nokogiri directly to soap wrapping
The prior approach involved to_xml which output a string that was then
parsed back into XML. This put out a processing instruction that led to
a parsing error. It's possible to omit the <?xml?> parsing instruction
on to_xml, but let's just pass a Node right to the builder.