Death to whitespace
Fixed a bunch of places where XML is serialized out to make sure it
never calls a plain to_xml since that pretty-prints the XML with
spaces. Will clean up to use a single method later, but this means the
SHA-1 checksums are now identical for both signed parts in the SHA-1 digests.