Successfully permits recreating a soap doc in Ruby to match Java
Expands upon the work Jacob has done to permit Ruby to recreate an
encrypted and signed SOAP document which [soon] matches that produced by
Java.
This also adds a spec which decrypts the message generated with Java;
however there is an error in the Ruby implementation we have not yet
conquered.