Manually implement xmlenc padding.
Implements http://www.w3.org/TR/xmlenc-core/#sec-Alg-Block padding since
openssl does not support this. See

  http://openssl.6102.n7.nabble.com/ISO10126-padding-in-openssl-td9228.html

for details on why openssl does not support it. Note that this padding method
is vulnerable to a very fast padding oracle attack

  https://www.nds.rub.de/media/nds/veroeffentlichungen/2011/10/22/HowToBreakXMLenc.pdf

but there is no alternative while remaining compatible with the xmlenc
spec. The mitigation is to use an authenticated cipher mode if the server
side supports it or to require a signature verification of the ciphertext
before attempting the decrypt to disallow tampering.