Handle transient errors (#227)
**Why**: There are certain types of errors we want to exclude from
Sentry. By defining those errors, and providing a `transient?` method,
it makes it easier for consumers of this gem to handle errors.

For example, instead of having logic in Caseflow to check if an error
came from BGS or VBMS, and whether or not it can be ignored, we can
move that logic to the `ruby-bgs` and `connect_vbms` gems, and
Caseflow now only needs to check if `error.transient?`.

This also eliminates the duplication of the same error handling logic
in both Caseflow and Efolder.