Allow raising custom errors
When we moved the custom errors from Caseflow to this gem, we forgot
to call `from_http_error` in `client.rb`, which meant that we were no
longer seeing the custom errors in Sentry, making it harder to figure
out which specific error was happening.

This PR also adds a few more ignorable errors to match what efolder
was doing.